11_03_2021 Security - Login/Logout Code Refactor - moving all components to a Components folder
