+++ ecommerce-project/frontend/angular-ecommerce/src/app/app.module.ts
@@ -6,14 +6,15 @@ import {ProductListComponent} from './components/product-list/product-list.compo
 import {HttpClientModule} from '@angular/common/http';
 import {ProductService} from './services/product.service';
 import {Routes, RouterModule} from '@angular/router';
-import {ProductCategoryMenuComponent} from './component/product-category-menu/product-category-menu.component';
+import {ProductCategoryMenuComponent} from './components/product-category-menu/product-category-menu.component';
 import {SearchComponent} from './components/search/search.component';
 import {ProductDetailsComponent} from './components/product-details/product-details.component';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
-import {CartStatusComponent} from './cart-status/cart-status.component';
-import {CartDetailsComponent} from './cart-details/cart-details.component';
-import {CheckoutComponent} from './checkout/checkout.component';
+import {CartStatusComponent} from './components/cart-status/cart-status.component';
+import {CartDetailsComponent} from './components/cart-details/cart-details.component';
+import {CheckoutComponent} from './components/checkout/checkout.component';
 import {ReactiveFormsModule} from '@angular/forms';
+import { LoginComponent } from './components/login/login.component';
 
 const routes: Routes = [
   {path: 'checkout', component: CheckoutComponent},
@@ -36,7 +37,8 @@ const routes: Routes = [
     ProductDetailsComponent,
     CartStatusComponent,
     CartDetailsComponent,
-    CheckoutComponent
+    CheckoutComponent,
+    LoginComponent
   ],
   imports: [
     RouterModule.forRoot(routes),
