+++ ecommerce-project/frontend/angular-ecommerce/src/app/components/cart-details/cart-details.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
-import {CartItem} from '../common/cart-item';
-import {CartService} from '../services/cart.service';
+import {CartItem} from '../../common/cart-item';
+import {CartService} from '../../services/cart.service';
 
 @Component({
   selector: 'app-cart-details',
