+++ ecommerce-project/frontend/angular-ecommerce/src/app/components/cart-status/cart-status.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {CartService} from '../services/cart.service';
+import {CartService} from '../../services/cart.service';
 
 @Component({
   selector: 'app-cart-status',
