+++ ecommerce-project/frontend/angular-ecommerce/src/app/components/checkout/checkout.component.html
