+++ ecommerce-project/frontend/angular-ecommerce/src/app/components/product-category-menu/product-category-menu.component.ts
